First attempts to remove test_rpcs.py boilerplate
